chore: clean up ranker ndcg typing conversion
Agent-Logs-Url: https://github.com/ashutosh2287/cricsmart/sessions/2b94abba-f243-4a52-8aa8-b16e110dc8de

+++ ml/commentary/training/train_template_ranker.py
@@ -109,7 +109,7 @@ def train_ranker(
     predictions = np.asarray(ranker.predict(transformed), dtype=float)
     ndcg = float(
         ndcg_score(
-            np.asarray([labels.to_numpy(dtype=float)], dtype=float),
+            np.asarray([labels.to_numpy(dtype=float)]),
             np.asarray([predictions], dtype=float),
         )
     )
